requirements.txt
gradio
openai
anthropic
supabase
python-dotenv
